Added Feedback mailing and manual concept tools mail button (#61)
Commit log:
* Added feedback notification emails

* WIP - Added send concept tools mail button to admin portal

* Fixed send concept tools mail button and various other things

* Fixed auth middleware being called at the wrong time

* Fixed concept tools email tools counter being off

* Fixed mail exception on travis by disabling mailing on traivs

Notes:
Mailing on Travis is currently not possible since we would have to create a paid account on MailTrap with a company email.
Scheduled mailing on Heroku is also not possible. Because we would need a seperate dyno to run 100% on the time with a Laravel app running that schedules the mailing and Heroku has a finite amount of dyno hours.